refactor ( component-educational-staff and componente - students

+++ src/app/pages/all-books/all-books.component.ts
@@ -4,7 +4,7 @@ import {MatTableDataSource} from '@angular/material/table';
 import { BooksService } from 'src/app/shared/services/books.service';
 import { Router } from '@angular/router';
 import { Book } from '../books/domain/book';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-all-books',
   templateUrl: './all-books.component.html',
@@ -43,6 +43,19 @@ export class AllBooksComponent implements OnInit {
     this.router.navigate(['/home/add-book']);
   }
   delete(book: any) {
-    this._service.delete(book.id).subscribe();
+
+    Swal.fire({
+      title: 'Esta seguro eliminar el elñemento?',
+      text: "}",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si!'
+    }).then((result) => {
+      if (result.value) {
+        this._service.delete(book.id).subscribe();
+      }
+    })
   }
 }
